+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/records/RecordCreationSuite.java
@@ -12,14 +12,17 @@ import static com.hedera.services.bdd.spec.assertions.TransferListAsserts.includ
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountRecords;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
 import static com.hedera.services.bdd.spec.transactions.TxnUtils.asId;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.createTopic;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.submitMessageTo;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uncheckedSubmit;
 import static com.hedera.services.bdd.spec.transactions.crypto.HapiCryptoTransfer.tinyBarsFromTo;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingHbar;
 import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.doWithStartupConfig;
+import static com.hedera.services.bdd.spec.utilops.UtilVerbs.assertionsHold;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
+import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.usableTxnIdNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hedera.services.bdd.suites.HapiSuite.FUNDING;
@@ -36,7 +39,10 @@ import com.hedera.services.bdd.junit.HapiTestLifecycle;
 import com.hedera.services.bdd.junit.LeakyHapiTest;
 import com.hedera.services.bdd.junit.LeakyRepeatableHapiTest;
 import com.hedera.services.bdd.junit.support.TestLifecycle;
+import com.hedera.services.bdd.spec.transactions.TxnUtils;
 import com.hederahashgraph.api.proto.java.AccountAmount;
+import com.hederahashgraph.api.proto.java.AccountID;
+import com.hederahashgraph.api.proto.java.TransactionRecord;
 import com.hederahashgraph.api.proto.java.TransferList;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.List;
@@ -81,36 +87,26 @@ public class RecordCreationSuite {
                         .memo(comfortingMemo)
                         .exposingFeesTo(feeObs)
                         .payingWith(PAYER),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) {
-                        feeObs.set(new FeeObject(16666, 150000, 0));
-                        // this wants the transaction to fail due to lack of funds, but should still pay the node and
-                        // network fee.
-                        // the problem is that the service fee is zero, so if they don't have enough for the whole thing
-                        // they don't
-                        // have enough for the node and network fee.  to fix this we make the transfer bigger
-                        // so the service fee is non zero, by including an extra account.
-                        return cryptoTransfer(
-                                        movingHbar(1L).between(GENESIS, FUNDING),
-                                        movingHbar(1L).between(GENESIS, TO_ACCOUNT))
-                                .memo(comfortingMemo)
-                                .fee(feeObs.get().networkFee() + feeObs.get().nodeFee())
-                                .payingWith(PAYER)
-                                .via(TXN_ID)
-                                .hasKnownStatus(INSUFFICIENT_TX_FEE)
-                                .logged();
-                    } else {
-                        return cryptoTransfer(tinyBarsFromTo(GENESIS, FUNDING, 1L))
-                                .memo(comfortingMemo)
-                                .fee(feeObs.get().networkFee() + feeObs.get().nodeFee())
-                                .payingWith(PAYER)
-                                .via(TXN_ID)
-                                .hasKnownStatus(INSUFFICIENT_TX_FEE)
-                                .logged();
-                    }
+                sourcing(() -> {
+                    feeObs.set(new FeeObject(16666, 150000, 0));
+                    // this wants the transaction to fail due to lack of funds, but should still pay the node and
+                    // network fee.
+                    // the problem is that the service fee is zero, so if they don't have enough for the whole thing
+                    // they don't
+                    // have enough for the node and network fee.  to fix this we make the transfer bigger
+                    // so the service fee is non zero, by including an extra account.
+                    return cryptoTransfer(
+                                    movingHbar(1L).between(GENESIS, FUNDING),
+                                    movingHbar(1L).between(GENESIS, TO_ACCOUNT))
+                            .memo(comfortingMemo)
+                            .fee(feeObs.get().networkFee() + feeObs.get().nodeFee())
+                            .payingWith(PAYER)
+                            .via(TXN_ID)
+                            .hasKnownStatus(INSUFFICIENT_TX_FEE)
+                            .logged();
                 }),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) feeObs.set(new FeeObject(16666, 150000, 0));
+                sourcing(() -> {
+                    feeObs.set(new FeeObject(16666, 150000, 0));
                     return getTxnRecord(TXN_ID)
                             .assertingNothingAboutHashes()
                             .hasPriority(recordWith()
@@ -121,7 +117,7 @@ public class RecordCreationSuite {
                                     .transfers(including(spec -> {
                                         final var networkFee = feeObs.get().networkFee();
                                         final var nodeFee = feeObs.get().nodeFee();
-                                        var fudge = "true".equals(flag) ? 0 : 1;
+                                        var fudge = 0;
                                         return TransferList.newBuilder()
                                                 .addAllAccountAmounts(List.of(
                                                         AccountAmount.newBuilder()
@@ -222,8 +218,8 @@ public class RecordCreationSuite {
                         .exposingFeesTo(feeObs)
                         .payingWith(PAYER),
                 usableTxnIdNamed(TXN_ID).payerId(PAYER),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) feeObs.set(new FeeObject(100000, 150000, 0));
+                sourcing(() -> {
+                    feeObs.set(new FeeObject(100000, 150000, 0));
                     return uncheckedSubmit(cryptoTransfer(tinyBarsFromTo(GENESIS, FUNDING, 1L))
                                     .memo(comfortingMemo)
                                     .fee(feeObs.get().networkFee() - 1L)
@@ -231,8 +227,8 @@ public class RecordCreationSuite {
                                     .txnId(TXN_ID))
                             .payingWith(GENESIS);
                 }),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) feeObs.set(new FeeObject(100000, 150000, 0));
+                sourcing(() -> {
+                    feeObs.set(new FeeObject(100000, 150000, 0));
                     return getTxnRecord(TXN_ID)
                             .assertingNothingAboutHashes()
                             .hasPriority(recordWith()
@@ -240,7 +236,7 @@ public class RecordCreationSuite {
                                             () -> 3L, feeObs.get().networkFee()))
                                     .transfers(including(spec -> {
                                         final var networkFee = feeObs.get().networkFee();
-                                        var fudge = "true".equals(flag) ? 0 : 1;
+                                        var fudge = 0;
                                         return TransferList.newBuilder()
                                                 .addAllAccountAmounts(List.of(
                                                         AccountAmount.newBuilder()
@@ -276,5 +272,33 @@ public class RecordCreationSuite {
                         .payingWith(PAYER)
                         .via(txn),
                 getAccountRecords(PAYER).has(inOrder(recordWith().txnId(txn))));
+    }
+
+    @HapiTest
+    final Stream<DynamicTest> payerRecordCreationSanityChecks() {
+        return hapiTest(
+                cryptoCreate(PAYER),
+                createTopic("ofGeneralInterest").payingWith(PAYER),
+                cryptoTransfer(tinyBarsFromTo(GENESIS, FUNDING, 1_000L)).payingWith(PAYER),
+                submitMessageTo("ofGeneralInterest").message("I say!").payingWith(PAYER),
+                assertionsHold((spec, opLog) -> {
+                    final var payerId = spec.registry().getAccountID(PAYER);
+                    final var subOp = getAccountRecords(PAYER).logged();
+                    allRunFor(spec, subOp);
+                    final var records = subOp.getResponse().getCryptoGetAccountRecords().getRecordsList().stream()
+                            .filter(TxnUtils::isNotEndOfStakingPeriodRecord)
+                            .toList();
+                    assertEquals(3, records.size());
+                    for (var record : records) {
+                        assertEquals(record.getTransactionFee(), -netChangeIn(record, payerId));
+                    }
+                }));
+    }
+
+    private long netChangeIn(TransactionRecord record, AccountID id) {
+        return record.getTransferList().getAccountAmountsList().stream()
+                .filter(aa -> id.equals(aa.getAccountID()))
+                .mapToLong(AccountAmount::getAmount)
+                .sum();
     }
 }
