+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/regression/TargetNetworkPrep.java
@@ -15,7 +15,6 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.crypto.HapiCryptoTransfer.tinyBarsFromTo;
 import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.balanceSnapshot;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.doWithStartupConfig;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.inParallel;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overridingTwo;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
@@ -26,13 +25,11 @@ import static com.hedera.services.bdd.suites.HapiSuite.NODE_REWARD;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HBAR;
 import static com.hedera.services.bdd.suites.HapiSuite.STAKING_REWARD;
 
-import com.hedera.node.app.hapi.utils.fee.FeeObject;
 import com.hedera.services.bdd.junit.LeakyEmbeddedHapiTest;
 import com.hedera.services.bdd.spec.HapiSpecOperation;
 import com.hedera.services.bdd.spec.assertions.AccountInfoAsserts;
 import com.hederahashgraph.api.proto.java.Key;
 import com.hederahashgraph.api.proto.java.KeyList;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DynamicTest;
@@ -49,7 +46,6 @@ public class TargetNetworkPrep {
                 Key.newBuilder().setKeyList(KeyList.getDefaultInstance()).build();
         final var snapshot802 = "802startBalance";
         final var civilian = "civilian";
-        final AtomicReference<FeeObject> feeObs = new AtomicReference<>();
         return hapiTest(
                 overridingTwo(
                         "nodes.feeCollectionAccountEnabled", "false", "nodes.preserveMinNodeRewardBalance", "false"),
@@ -58,30 +54,15 @@ public class TargetNetworkPrep {
                 cryptoTransfer(tinyBarsFromTo(civilian, STAKING_REWARD, ONE_HBAR))
                         .payingWith(civilian)
                         .signedBy(civilian)
-                        .exposingFeesTo(feeObs)
                         .via("transferToStakingReward"),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) {
-                        return getTxnRecord("transferToStakingReward")
-                                .hasHbarAmount(STAKING_REWARD, (long) (ONE_HBAR + simpleNetworkAndServiceFee * 0.1));
-                    } else {
-                        return getTxnRecord("transferToStakingReward").hasHbarAmount(STAKING_REWARD, (long) (ONE_HBAR
-                                + ((feeObs.get().networkFee() + feeObs.get().serviceFee()) * 0.1)));
-                    }
-                }),
+                getTxnRecord("transferToStakingReward")
+                        .hasHbarAmount(STAKING_REWARD, (long) (ONE_HBAR + simpleNetworkAndServiceFee * 0.1)),
                 cryptoTransfer(tinyBarsFromTo(civilian, NODE_REWARD, ONE_HBAR))
                         .payingWith(civilian)
                         .signedBy(civilian)
                         .via("transferToNodeReward"),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) {
-                        return getTxnRecord("transferToNodeReward")
-                                .hasHbarAmount(NODE_REWARD, (long) (ONE_HBAR + simpleNetworkAndServiceFee * 0.1));
-                    } else {
-                        return getTxnRecord("transferToNodeReward").hasHbarAmount(NODE_REWARD, (long) (ONE_HBAR
-                                + ((feeObs.get().networkFee() + feeObs.get().serviceFee()) * 0.1)));
-                    }
-                }),
+                getTxnRecord("transferToNodeReward")
+                        .hasHbarAmount(NODE_REWARD, (long) (ONE_HBAR + simpleNetworkAndServiceFee * 0.1)),
                 getAccountDetails(STAKING_REWARD)
                         .payingWith(GENESIS)
                         .has(accountDetailsWith()
@@ -132,7 +113,6 @@ public class TargetNetworkPrep {
         final var emptyKey =
                 Key.newBuilder().setKeyList(KeyList.getDefaultInstance()).build();
         final var civilian = "civilian";
-        final AtomicReference<FeeObject> feeObs = new AtomicReference<>();
         return hapiTest(
                 overridingTwo(
                         "nodes.feeCollectionAccountEnabled", "true", "nodes.preserveMinNodeRewardBalance", "false"),
@@ -140,34 +120,17 @@ public class TargetNetworkPrep {
                 cryptoTransfer(tinyBarsFromTo(civilian, STAKING_REWARD, ONE_HBAR))
                         .payingWith(civilian)
                         .signedBy(civilian)
-                        .exposingFeesTo(feeObs)
                         .via("stakingRewardTransfer"),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) {
-                        return getTxnRecord("stakingRewardTransfer")
-                                .hasHbarAmount(STAKING_REWARD, ONE_HBAR)
-                                .hasHbarAmount(FEE_COLLECTOR, simpleCryptoTransferFee);
-                    } else {
-                        return getTxnRecord("stakingRewardTransfer")
-                                .hasHbarAmount(STAKING_REWARD, ONE_HBAR)
-                                .hasHbarAmount(FEE_COLLECTOR, feeObs.get().totalFee());
-                    }
-                }),
+                getTxnRecord("stakingRewardTransfer")
+                        .hasHbarAmount(STAKING_REWARD, ONE_HBAR)
+                        .hasHbarAmount(FEE_COLLECTOR, simpleCryptoTransferFee),
                 cryptoTransfer(tinyBarsFromTo(civilian, NODE_REWARD, ONE_HBAR))
                         .payingWith(civilian)
                         .signedBy(civilian)
                         .via("nodeRewardTransfer"),
-                doWithStartupConfig("fees.simpleFeesEnabled", flag -> {
-                    if ("true".equals(flag)) {
-                        return getTxnRecord("nodeRewardTransfer")
-                                .hasHbarAmount(NODE_REWARD, ONE_HBAR)
-                                .hasHbarAmount(FEE_COLLECTOR, simpleCryptoTransferFee);
-                    } else {
-                        return getTxnRecord("nodeRewardTransfer")
-                                .hasHbarAmount(NODE_REWARD, ONE_HBAR)
-                                .hasHbarAmount(FEE_COLLECTOR, feeObs.get().totalFee());
-                    }
-                }),
+                getTxnRecord("nodeRewardTransfer")
+                        .hasHbarAmount(NODE_REWARD, ONE_HBAR)
+                        .hasHbarAmount(FEE_COLLECTOR, simpleCryptoTransferFee),
                 getAccountDetails(STAKING_REWARD)
                         .payingWith(GENESIS)
                         .has(accountDetailsWith()
